requirements.txt
great_expectations==0.18.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
# pyarrow==14.0.1
fastparquet==2023.10.1
dagster==1.5.9
apache-beam==2.52.0
duckdb==0.9.2
